Create multiple extension projects
