+++ GitDiffMargin.Extension.15/GitDiffMargin.Extension.15.csproj
@@ -2,16 +2,21 @@
 <Project ToolsVersion="15.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
   <PropertyGroup>
+    <MinimumVisualStudioVersion>15.0</MinimumVisualStudioVersion>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{AFEE8D6F-0358-4628-8CAC-D10668C95057}</ProjectGuid>
+    <ProjectTypeGuids>{82b43b9b-a64c-4715-b499-d71e9ca2bd60};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
+    <ProjectGuid>{FBDDDD6F-2CE8-446B-AA3F-87236A2158F4}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>GitDiffMargin.Commands</RootNamespace>
-    <AssemblyName>GitDiffMargin.Commands</AssemblyName>
+    <RootNamespace>GitDiffMargin.Extension</RootNamespace>
+    <AssemblyName>GitDiffMargin.Extension</AssemblyName>
     <TargetFrameworkVersion>v4.6</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
-    <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
+    <GeneratePkgDefFile>true</GeneratePkgDefFile>
+    <IncludeAssemblyInVSIXContainer>false</IncludeAssemblyInVSIXContainer>
+    <IncludeDebugSymbolsInLocalVSIXDeployment>false</IncludeDebugSymbolsInLocalVSIXDeployment>
+    <TargetVsixContainerName>GitDiffMargin.15.vsix</TargetVsixContainerName>
     <TargetFrameworkProfile />
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
@@ -32,8 +37,14 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <PropertyGroup>
-    <SignAssembly>true</SignAssembly>
-    <AssemblyOriginatorKeyFile>..\GitDiffMargin\SharedKey.snk</AssemblyOriginatorKeyFile>
+    <!-- Common debugging support -->
+    <StartAction>Program</StartAction>
+    <StartProgram>$(DevEnvDir)\devenv.exe</StartProgram>
+    <StartArguments>/rootSuffix Exp</StartArguments>
+  </PropertyGroup>
+  <PropertyGroup Condition="'$(BuildingInsideVisualStudio)' != 'true'">
+    <!-- This property disables extension deployment for command line builds; required for AppVeyor -->
+    <DeployExtension>False</DeployExtension>
   </PropertyGroup>
   <ItemGroup>
     <Reference Include="System" />
@@ -44,36 +55,23 @@
     <PackageReference Include="Microsoft.VisualStudio.Threading" Version="15.6.56" />
     <PackageReference Include="Newtonsoft.Json" Version="9.0.1" />
     <PackageReference Include="StreamJsonRpc" Version="1.3.23" />
+    <PackageReference Include="Microsoft.VSSDK.BuildTools" Version="15.5.72" PrivateAssets="all" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="CommandHandlerTextViewCreationListener.cs" />
-    <Compile Include="CopyOldTextCommandArgs.cs" />
-    <Compile Include="CopyOldTextCommandHandler.cs" />
-    <Compile Include="GitDiffMarginCommandBinding.cs" />
-    <Compile Include="GitDiffMarginCommandHandler`1.cs" />
-    <Compile Include="NextChangeCommandArgs.cs" />
-    <Compile Include="NextChangeCommandHandler.cs" />
-    <Compile Include="PreviousChangeCommandArgs.cs" />
-    <Compile Include="PreviousChangeCommandHandler.cs" />
-    <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="RollbackChangeCommandArgs.cs" />
-    <Compile Include="RollbackChangeCommandHandler.cs" />
-    <Compile Include="ShimCommandHandler.cs" />
-    <Compile Include="ShowDiffCommandArgs.cs" />
-    <Compile Include="ShowDiffCommandHandler.cs" />
-    <Compile Include="ShowPopupCommandArgs.cs" />
-    <Compile Include="ShowPopupCommandHandler.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\GitDiffMargin\GitDiffMargin.csproj">
+    <ProjectReference Include="..\GitDiffMargin\GitDiffMargin.15.csproj">
       <Project>{233b6f13-7687-4823-870b-a8f44b79d4f0}</Project>
-      <Name>GitDiffMargin</Name>
+      <Name>GitDiffMargin.15</Name>
+      <IncludeOutputGroupsInVSIX>BuiltProjectOutputGroup;BuiltProjectOutputGroupDependencies;GetCopyToOutputDirectoryItems;SatelliteDllsProjectOutputGroup;PkgDefProjectOutputGroup</IncludeOutputGroupsInVSIX>
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
-    <None Include="..\GitDiffMargin\SharedKey.snk">
-      <Link>SharedKey.snk</Link>
+    <Compile Include="Properties\AssemblyInfo.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="source.extension.vsixmanifest">
+      <SubType>Designer</SubType>
     </None>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
+  <Import Project="$(VSToolsPath)\VSSDK\Microsoft.VsSDK.targets" Condition="Exists('$(VSToolsPath)\VSSDK\Microsoft.VsSDK.targets')" />
 </Project>
