+++ GitDiffMargin.Extension.16/source.extension.vsixmanifest
@@ -13,18 +13,15 @@
         <!--<Tags></Tags>-->
     </Metadata>
     <Installation>
-        <InstallationTarget Version="[11.0,18.0)" Id="Microsoft.VisualStudio.Community" />
-        <InstallationTarget Version="[11.0,18.0)" Id="Microsoft.VisualStudio.Pro" />
+        <InstallationTarget Version="[15.0.27428,17.0)" Id="Microsoft.VisualStudio.Community" />
+        <InstallationTarget Version="[15.0.27428,17.0)" Id="Microsoft.VisualStudio.Pro" />
     </Installation>
     <Prerequisites>
-        <Prerequisite Id="Microsoft.VisualStudio.Component.CoreEditor" Version="[15.0,18.0)" DisplayName="Visual Studio core editor" />
+        <Prerequisite Id="Microsoft.VisualStudio.Component.CoreEditor" Version="[15.0.27428,17.0)" DisplayName="Visual Studio core editor" />
     </Prerequisites>
     <Assets>
-        <Asset Type="Microsoft.VisualStudio.MefComponent" d:Source="Project" d:ProjectName="GitDiffMargin.15" TargetVersion="[15.0,15.0.27131)" Path="|GitDiffMargin.15|" />
-        <Asset Type="Microsoft.VisualStudio.MefComponent" d:Source="Project" d:ProjectName="GitDiffMargin.16" TargetVersion="[15.0.27428,17.0)" Path="|GitDiffMargin.16|" />
-        <Asset Type="Microsoft.VisualStudio.MefComponent" d:Source="Project" d:ProjectName="GitDiffMargin.Commands" Path="|GitDiffMargin.Commands|" />
-        <Asset Type="Microsoft.VisualStudio.VsPackage" d:Source="Project" d:ProjectName="GitDiffMargin.15" TargetVersion="[15.0,15.0.27131)" Path="|GitDiffMargin.15;PkgdefProjectOutputGroup|" />
-        <Asset Type="Microsoft.VisualStudio.VsPackage" d:Source="Project" d:ProjectName="GitDiffMargin.16" TargetVersion="[15.0.27428,17.0)" Path="|GitDiffMargin.16;PkgdefProjectOutputGroup|" />
+        <Asset Type="Microsoft.VisualStudio.MefComponent" d:Source="Project" d:ProjectName="GitDiffMargin.16" Path="|GitDiffMargin.16|" />
+        <Asset Type="Microsoft.VisualStudio.VsPackage" d:Source="Project" d:ProjectName="GitDiffMargin.16" Path="|GitDiffMargin.16;PkgdefProjectOutputGroup|" />
         <Asset Type="Microsoft.VisualStudio.VsPackage" d:Source="Project" d:ProjectName="%CurrentProject%" Path="|%CurrentProject%;PkgdefProjectOutputGroup|" />
     </Assets>
 </PackageManifest>
