+++ GitDiffMargin.Shared/Commands/CopyOldTextCommandArgs.cs
@@ -1,14 +1,18 @@
-﻿namespace GitDiffMargin.Commands
+﻿#if !LEGACY_COMMANDS
+
+namespace GitDiffMargin.Commands
 {
     using Microsoft.VisualStudio.Text;
     using Microsoft.VisualStudio.Text.Editor;
     using Microsoft.VisualStudio.Text.Editor.Commanding;
 
-    internal class ShowPopupCommandArgs : EditorCommandArgs
+    internal class CopyOldTextCommandArgs : EditorCommandArgs
     {
-        public ShowPopupCommandArgs(ITextView textView, ITextBuffer subjectBuffer)
+        public CopyOldTextCommandArgs(ITextView textView, ITextBuffer subjectBuffer)
             : base(textView, subjectBuffer)
         {
         }
     }
 }
+
+#endif
