+++ GitDiffMargin.Shared/Commands/RollbackChangeCommandArgs.cs
@@ -1,14 +1,18 @@
-﻿namespace GitDiffMargin.Commands
+﻿#if !LEGACY_COMMANDS
+
+namespace GitDiffMargin.Commands
 {
     using Microsoft.VisualStudio.Text;
     using Microsoft.VisualStudio.Text.Editor;
     using Microsoft.VisualStudio.Text.Editor.Commanding;
 
-    internal class PreviousChangeCommandArgs : EditorCommandArgs
+    internal class RollbackChangeCommandArgs : EditorCommandArgs
     {
-        public PreviousChangeCommandArgs(ITextView textView, ITextBuffer subjectBuffer)
+        public RollbackChangeCommandArgs(ITextView textView, ITextBuffer subjectBuffer)
             : base(textView, subjectBuffer)
         {
         }
     }
 }
+
+#endif
