+++ GitDiffMargin.Shared/Commands/ShowPopupCommandArgs.cs
@@ -1,14 +1,18 @@
-﻿namespace GitDiffMargin.Commands
+﻿#if !LEGACY_COMMANDS
+
+namespace GitDiffMargin.Commands
 {
     using Microsoft.VisualStudio.Text;
     using Microsoft.VisualStudio.Text.Editor;
     using Microsoft.VisualStudio.Text.Editor.Commanding;
 
-    internal class RollbackChangeCommandArgs : EditorCommandArgs
+    internal class ShowPopupCommandArgs : EditorCommandArgs
     {
-        public RollbackChangeCommandArgs(ITextView textView, ITextBuffer subjectBuffer)
+        public ShowPopupCommandArgs(ITextView textView, ITextBuffer subjectBuffer)
             : base(textView, subjectBuffer)
         {
         }
     }
 }
+
+#endif
