+++ GitDiffMargin.Shared/LegacyCommands/LegacyCommandHandlerTextViewCreationListener.cs
@@ -1,4 +1,6 @@
-﻿namespace GitDiffMargin.Commands
+﻿#if LEGACY_COMMANDS
+
+namespace GitDiffMargin.LegacyCommands
 {
     using System.ComponentModel.Composition;
     using Microsoft.VisualStudio.Editor;
@@ -10,13 +12,13 @@
     [Export(typeof(IVsTextViewCreationListener))]
     [ContentType("text")]
     [TextViewRole(PredefinedTextViewRoles.Editable)]
-    internal class CommandHandlerTextViewCreationListener : IVsTextViewCreationListener
+    internal class LegacyCommandHandlerTextViewCreationListener : IVsTextViewCreationListener
     {
         private readonly SVsServiceProvider _serviceProvider;
         private readonly IVsEditorAdaptersFactoryService _editorAdaptersFactoryService;
 
         [ImportingConstructor]
-        public CommandHandlerTextViewCreationListener(SVsServiceProvider serviceProvider, IVsEditorAdaptersFactoryService editorAdaptersFactoryService)
+        public LegacyCommandHandlerTextViewCreationListener(SVsServiceProvider serviceProvider, IVsEditorAdaptersFactoryService editorAdaptersFactoryService)
         {
             _serviceProvider = serviceProvider;
             _editorAdaptersFactoryService = editorAdaptersFactoryService;
@@ -28,10 +30,11 @@
             if (textView == null)
                 return;
 
-            // The new command handling approach does not require that the command filter be enabled. The command
-            // implementations interact directly with the handler via its IOleCommandTarget interface.
             GitDiffMarginCommandHandler filter = new GitDiffMarginCommandHandler(textViewAdapter, _editorAdaptersFactoryService, textView);
+            filter.Enabled = true;
             textView.Properties.AddProperty(typeof(GitDiffMarginCommandHandler), filter);
         }
     }
 }
+
+#endif
