+++ GitDiffMargin/GitDiffMargin.15.csproj
@@ -1,5 +1,10 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="12.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup>
+    <MSBuildProjectExtensionsPath>obj\15\</MSBuildProjectExtensionsPath>
+    <BaseIntermediateOutputPath>obj\15\</BaseIntermediateOutputPath>
+  </PropertyGroup>
+
   <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
   <Import Project="$(PkgLibGit2Sharp_NativeBinaries)\build\LibGit2Sharp.NativeBinaries.props" Condition="Exists('$(PkgLibGit2Sharp_NativeBinaries)\build\LibGit2Sharp.NativeBinaries.props')" />
   <PropertyGroup>
@@ -13,7 +18,7 @@
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>GitDiffMargin</RootNamespace>
-    <AssemblyName>GitDiffMargin</AssemblyName>
+    <AssemblyName>GitDiffMargin.15</AssemblyName>
     <TargetFrameworkVersion>v4.6</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <UseCodebase>true</UseCodebase>
@@ -28,8 +33,8 @@
     <DebugSymbols>true</DebugSymbols>
     <DebugType>full</DebugType>
     <Optimize>false</Optimize>
-    <OutputPath>bin\Debug\</OutputPath>
-    <DefineConstants>DEBUG;TRACE</DefineConstants>
+    <OutputPath>bin\15\$(Configuration)\</OutputPath>
+    <DefineConstants>DEBUG;TRACE;LEGACY_COMMANDS</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
     <Prefer32Bit>false</Prefer32Bit>
@@ -37,8 +42,8 @@
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
     <DebugType>pdbonly</DebugType>
     <Optimize>true</Optimize>
-    <OutputPath>bin\Release\</OutputPath>
-    <DefineConstants>TRACE</DefineConstants>
+    <OutputPath>bin\15\$(Configuration)\</OutputPath>
+    <DefineConstants>TRACE;LEGACY_COMMANDS</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
     <Prefer32Bit>false</Prefer32Bit>
